Hack solution for furo theme
The main issue is that the asciinema-player requires the js library to loaded before the <script> tag is read.
To optimize the page-load speed, furo loads all of the js files at the end of the body tag (similar to calling <script defer>).
In this scenario, the asciinema-player js file would be loaded to late and the asciinema video cannot be inserted, since the function to load the player is not yet defined.
A hacky workaround is to simply include the js file before inserting the js-code to insert the video.
However, this is a blocking operation and requires the browser to first download the js library...
So a future version should correctly fix this issue.

Another issue that was found is that a transform is applied to **all** documents.
The current implementation has terrible performance because the recordings are generated multiple times!